requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.6
    # via fastapi-cache2
aiosignal==1.3.1
    # via aiohttp
alembic==1.12.0
    # via -r requirements.in
amqp==5.1.1
    # via kombu
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via
    #   aiohttp
    #   redis
attrs==23.1.0
    # via aiohttp
billiard==4.1.0
    # via celery
celery==5.3.4
    # via
    #   -r requirements.in
    #   flower
charset-normalizer==3.3.0
    # via aiohttp
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.4.2
    # via email-validator
email-validator==2.0.0.post2
    # via pydantic
fastapi==0.103.2
    # via
    #   -r requirements.in
    #   fastapi-cache2
fastapi-cache2[redis]==0.2.1
    # via -r requirements.in
flower==2.0.1
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.0
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.0
    # via uvicorn
humanize==4.8.0
    # via flower
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   yarl
kombu==5.3.2
    # via celery
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via mako
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
passlib==1.7.4
    # via -r requirements.in
pendulum==2.1.2
    # via fastapi-cache2
prometheus-client==0.17.1
    # via flower
prompt-toolkit==3.0.39
    # via click-repl
psycopg2==2.9.9
    # via -r requirements.in
pydantic[email]==2.4.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.10.1
    # via pydantic
python-dateutil==2.8.2
    # via
    #   celery
    #   pendulum
python-dotenv==1.0.0
    # via uvicorn
pytz==2023.3.post1
    # via flower
pytzdata==2020.1
    # via pendulum
pyyaml==6.0.1
    # via uvicorn
redis==4.6.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.21
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
tornado==6.3.3
    # via flower
typing-extensions==4.8.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2023.3
    # via celery
uvicorn[standard]==0.23.2
    # via
    #   -r requirements.in
    #   fastapi-cache2
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.20.0
    # via uvicorn
wcwidth==0.2.8
    # via prompt-toolkit
websockets==11.0.3
    # via uvicorn
yarl==1.9.2
    # via aiohttp
